feat: flask jwt extended

+++ flask-app/resources/item.py
@@ -1,27 +1,34 @@
-from flask import app, jsonify
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from db import db
 from schemas import ItemSchema, ItemUpdateSchema
 from models import ItemModel
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt
 
 bp = Blueprint("items", __name__, description="Operations on items")
 
 
-@bp.route("/item/<string:item_id>")
+@bp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @bp.response(200, ItemSchema)
     def get(self, item_id):
         return ItemModel.query.get_or_404(item_id)
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        print(jwt)
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
 
         return {"message": "Item deleted."}
 
+    @jwt_required()
     @bp.arguments(ItemUpdateSchema)
     @bp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -41,10 +48,12 @@ class Item(MethodView):
 
 @bp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @bp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
     @bp.arguments(
         ItemSchema,
     )
